+++ jquery.stl/jquery.stl.priority_queue.js
@@ -20,7 +20,7 @@
 
         var uniqueMembers = {};
 
-        var instance = $.extend({}, containerMembers, uniqueMembers);
+        var instance = $.extend({type: 'priority_queue'}, containerMembers, uniqueMembers);
         instance.new(container);
         return instance;
     };