+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+0.3b
+---
+- Sequence Containers
+    vector
+    deque
+    list *
+
+- Associative Containers
+    set *
+    multiset *
+    map *
+    multimap *
+
+- Other Containers
+    bitset *
+
+- Container Adaptors
+    stack
+    queue
+    priority_queue *
+
+(*) - not fully functional
+
 0.2b
 ---
 - Container classes: vector, deque, list